PlayerMovement adjustment
+ changed PlayerMovement Script
+ changed Sliding script
+ changed ObjectsOnTrain Script

There is no sprinting now. The player always moves at sprint speed. Shift now instead slides you down ramps which works the same as it did before it just now is relocated  to shift due to ctrl not working well with browser hot keys.
Speed is now reset when landing on a train and movement in the air is slower but much more consistent as the force applied is always the same regardless of momentum built up. The movement should feel a bit easier and more consistent to the players input